Moved Player class to new package.
